Added /grass, and removed ability of non-ops to place grass blocks directly. Fixed server not reporting accurate player count to heartbeat. Assorted code cleanup and commenting.

+++ FemtoCraft/Config.cs
@@ -54,13 +54,13 @@ namespace FemtoCraft {
                             VerifyNames = Boolean.Parse( value );
                             break;
                         case "port":
-                            Port = Int32.Parse( value );
+                            Port = UInt16.Parse( value );
                             break;
                         case "admin-slot":
                             AdminSlot = Boolean.Parse( value );
                             break;
                         case "max-players":
-                            MaxPlayers = Int32.Parse( value );
+                            MaxPlayers = Byte.Parse( value );
                             break;
                         case "max-connections":
                             MaxConnections = Int32.Parse( value );
